data_processor/requirements.txt
pyspark
pandas
python-dotenv
sqlalchemy
pymysql
boto3